+++ changelogs/fragments/78156-undefined-check-in-finalize.yml
@@ -1,0 +1,2 @@
+bugfixes:
+  - Move undefined check from concat to finalize (https://github.com/ansible/ansible/issues/78156)
