+++ test/integration/targets/template/undefined_in_import.j2
@@ -1,0 +1,1 @@
+{% import 'undefined_in_import-import.j2' as t %}
